+++ lib/core.dart
@@ -2,9 +2,11 @@
 library core;
 
 export 'src/core/len.dart';
+export 'src/core/match.dart';
 export 'src/core/pair.dart';
 export 'src/core/req.dart';
 export 'src/core/req_len.dart';
+export 'src/core/req_match.dart';
 export 'src/core/rules.dart';
 export 'src/core/to_val_obj.dart';
 export 'src/core/trim.dart';